--- src/repositories/index.ts
@@ -1,1 +1,0 @@
-export { IUserRepository } from './user';
