+++ src/services/index.ts
@@ -1,2 +1,3 @@
 export { UserService } from './user';
+export { IUserRepository } from './user';
