--- src/services/interfaces/db-manager.ts
@@ -1,6 +1,0 @@
-import { BaseEntity } from 'typeorm';
-
-export interface IDbManager {
-  save: (model: BaseEntity) => Promise<BaseEntity>;
-  findOne: (table: typeof BaseEntity, id: number) => BaseEntity;
-}
