--- src/services/interfaces/index.ts
@@ -1,1 +1,0 @@
-export { IDbManager } from './db-manager';
