+++ src/services/user/index.ts
@@ -1,1 +1,2 @@
 export { UserService } from './class';
+export { IUserRepository } from './interfaces';
