+++ src/services/user/interfaces/index.ts
@@ -1,0 +1,1 @@
+export { IUserRepository } from './user-repository';
