+++ src/services/user/interfaces/user-repository.ts
@@ -1,7 +1,8 @@
-import { IUserDto, IUserModel } from '../../models';
+import { IUserDto, IUserModel } from '../../../models';
 
 export interface IUserRepository {
   create: (data: IUserDto) => IUserModel;
   findOne: (id: number) => Promise<IUserModel>;
   save: (user: IUserModel) => Promise<IUserModel>;
 }
+
